modules/api-location-service-gRPC/requirements.txt
grpcio
grpcio-tools
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
SQLAlchemy==1.3.19
GeoAlchemy2==0.8.4
shapely==1.7.0
python-dateutil==2.8.1
psycopg2-binary==2.8.5